refactor: unify ChunkProcessor to use StreamingAudioSampleSource
- Refactor ChunkProcessor to accept StreamingAudioSampleSource
- Add convenience init(audioSamples:) for backward compatibility
- Delete StreamingChunkProcessor (was duplicate code)
- Update transcribeStreaming() to use unified ChunkProcessor
